requirements.txt
selenium>=4.25.0
python-dotenv>=1.0.1
pyautogui>=0.9.54
flask>=3.0.0
cryptography>=41.0.0
markdown>=3.5.0
pyinstaller>=6.0.0

